backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
requests==2.31.0
